Add files via upload
IDLE Script Mod 7 Stephanie Hopkins
Create a list of 15 numbers.  You will use a loop to go through your list, and if the number is odd, you will print "the number is even" or "the number is odd," where the words the number should be replaced with the number in your list.  Example:  1 is odd -  2 is even